+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
  - Only require `bisect_ppx` and `alcotest` for testing, not for production builds.
+ - Check integers according to their bit lengths (issue https://github.com/hirrolot/mazeppa/issues/32).
 
 ## 0.7.0 - 2025-11-22